Cleaning up initial User tests.
